+++ source/Private/Get-nxEscapedPath.ps1
@@ -1,0 +1,13 @@
+function Get-nxEscapedPath
+{
+    [CmdletBinding()]
+    [OutputType([String])]
+    param
+    (
+        [Parameter()]
+        [System.String]
+        $Path
+    )
+
+    return ('"{0}"' -f $Path)
+}