+++ source/Public/usersAndGroups/Remove-nxLocalGroupMember.ps1
@@ -1,0 +1,58 @@
+function Remove-nxLocalGroupMember
+{
+    [CmdletBinding(SupportsShouldProcess = $true)]
+    [OutputType([void])]
+    param (
+        [Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true)]
+        [String[]]
+        $UserName,
+
+        [Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true)]
+        [String]
+        $GroupName,
+
+        [Parameter(ValueFromPipelineByPropertyName = $true)]
+        [Switch]
+        $PassThru
+    )
+
+    begin
+    {
+        $verbose = ($PSBoundParameters.ContainsKey('Verbose') -and $PSBoundParameters['Verbose']) -or $VerbosePreference -ne 'SilentlyContinue'
+    }
+
+    process
+    {
+        $hasGroupChanged = $false
+
+        foreach ($UserNameItem in $UserName)
+        {
+            $gpasswdParams = @('-d', $UserNameItem, $GroupName)
+            if ($PSCmdlet.ShouldProcess(
+                    "Performing the unix command 'gpasswd $($gpasswdParams -join ' ')'.",
+                    $UserNameItem,
+                    "Removing $userNameItem grom group '$GroupName'."
+                )
+            )
+            {
+                Invoke-NativeCommand -Executable 'gpasswd' -Parameters $gpasswdParams -Verbose:$verbose -ErrorAction 'Stop' | ForEach-Object -Process {
+                    if ($_ -match '^gpasswd:')
+                    {
+                        throw $_
+                    }
+                    else
+                    {
+                        Write-Verbose -Message $_
+                    }
+                }
+
+                $hasGroupChanged = $true
+            }
+        }
+
+        if ($hasGroupChanged)
+        {
+            Get-nxLocalGroup -GroupName $GroupName
+        }
+    }
+}