+++ tests/integration/Get-LsbRelease.tests.ps1
@@ -1,4 +1,4 @@
-Describe 'Get-LsbRelease function' -skip:$IsWindows {
+Describe 'Get-LsbRelease function' -skip:($IsWindows -or $IsMacOS) {
     it 'should not throw (if the binary is installed)' {
      {Get-LsbRelease} | Should -not -Throw
     }